Adding get/set field attrs
- The `get`/`set` field attributes control get-/set-ability from outside
the type. A field with no attributes will default to being gettable and
settable; and field with both `get` and `set` will also be both gettable
and settable (though this will produce a warning in a future version).
Otherwise, a field with just `get` will _only_ be gettable (and not
settable), and a field with just `set` will _only_ be settable (and not
gettable).
  - This last example will also produce a warning in a future version,
  since it doesn't really make sense for a field to be settable but
  _not_ gettable. Also, an empty field attr list
  (`type A { a: Int { } }`) will also produce a warning in a future
  version.