Typechecking for imports and exports
- Build module loader arch using ModuleLoader trait. This allows for
fs-based module loading when running the CLI, in-memory module loading
when done in WASM (for the abra site, at least), and project-based
loading when used in the language server. Add additional checks for
module loadability (if there was an error loading the module) and also
for circular references.
- Rename and refactor the test functions so that the main top-level code
is used. This way, the ModuleLoader codepath is always hit, and tests
can be controlled from the outside via a mock implementation of the
ModuleLoader trait.