Compilation of imports and exports
- Tons of changes here, I'll start with `ModuleLoader`. The
`ModuleLoader` is now responsible for a bit more - it's the main
orchestrator for typechecking and compilation, keeping track of the
modules that it's typechecked and then compiling those in the necessary
dependency order.
- Compiled constants are also stored in the module loader; this is used
to facilitate imports for Types/Enums, since Type/Enum values can just be
statically referenced as constants and don't need to be loaded from
globals like other bindings.
- The VM now processes multiple modules and persists state across modules.
- Constant/GLoad/GStore opcodes now reference `module_idx` as well as
`constant_idx`, since the VM needs to know which "module" to load the
constant from. I could foresee this being refactored in the future.