Generic enums
- Support generics in enums. This work was mostly already implemented,
it turns out. I just needed to add the correct pieces to the correct
places and it's fine! I do think the whole `Type::Enum` vs
`Type::EnumVariant` disparity needs to be addressed at some point,
hopefully in a future PR. There doesn't seem to be a need for that level
of indirection