Function return value overhaul
- All functions now always return a value. Unit-returning functions will
return a `None` value implicitly, which will always be popped off at
runtime because an expression of type Unit is not assignable to any
variable. This may seem a bit wasteful, but it actually enables lots of
conveniences, like passing values to functions which expect `() =>
Unit`, but actually return some value, eg.
```
var count = 0
[1, 2, 3].forEach(i => count += i)
```
This lambda _returns_ the result of `count += i`, but that result is
popped when being evaluated.
- This also cleans up a lot of function return types internally, which
can now always deal in `Value` rather than `Option<Value>` (since
previously, a Unit-returning function would manifest as a `None`
internally), as well as tons of tests.
- This also adds the `Array#forEach` method, as a proof of concept of
this working correctly.