Writing array tests in native abra
- Fully flesh out the `array_test.abra` test suite with unit tests for
every field, operator, static method, and instance method for arrays.
- To support this, I found a bug related to the previous change; when a
`return` statement is used in a Unit-returning function, we need to
manually push a `Value::Nil` onto the stack so it'll be returned. Also,
let's just go ahead and make sure that Unit-returning lambdas _also_
work correctly.
- Add temporary code to the LSP module to get
`ParseError::UnexpectedEof` to work; this should be fixed in a future
changeset.