Large refactor of globals
- Previously, globals were kind of a mess. Within the VM, globals was a
hashmap from name->value, and the name was a `Value::Str` that needed to
be referenced in the runtime `constants` array. Now, we keep track of
globals during compilation and we're able to reference globals by index
at runtime. This saves a few steps during runtime (good!) and also
allows us to remove the `Value::Str` variant entirely (awesome!).
- This also cleans up a lot of things with `constants` as well; in
addition to no longer having many unnecessary `Value::Str` constant
values, native modules' imports are now able to be referenced by global
index rather than by constant index. This allows for the cleanup
of #306, and it also helps pave the way for garbage collection (as well
as supporting potentially mutable imports).
- Also, now that types are referenced by their global index, we can
streamline the instantiation process. The `Opcode::New` instruction
contains a direct reference to the global index corresponding to the
Type being instantiated, so this saves an instruction (previously we'd
need to push the global/constant onto the stack first).
- This _also_ simplifies construction of native types (as well as
`type_id` representations in general), since the `module_idx` is no
longer needed. We can also remove a lot of the code in the ModuleLoader,
since it doesn't need to be responsible for keeping track of as many
things anymore.