Remove a lot of the unnecessary wasm stuff
- When I initially went down the wasm rabbit hole, I went a little too
hard. We do not need js translations of all the runtime values, all the
types, all the tokens, all the errors, etc. Converting values and error
messages to strings should be more than good enough (especially
considering the primary use-case of the wasm lib at the moment is the
docs site).