Fixing returns
- Right now, returning from a block within a function does not properly
pop that scope's locals, leaving garbage on top of the stack. This is
because the `first_local_idx` value that's being used during
`visit_return` is not w.r.t. the function scope, but rather the
enclosing for/while/match/if scope. To fix this, we can actually just
always emit `LStore 0` to store the return valueinto the return local slot.