Add try-expressions!
- Add try-expressions. These take the form `try <expr>` and attempt to
"unwrap" the value of the expression to which it's applied (as a prefix
operator). If this value is some "successful" value, the function flow
proceeds as normal; if it's some "failure" value, the failure value is
returned and the function terminates. (This is very similar to, and
largely inspired by, how Rust's `?` operator works).
- After typechecking is performed to ensure the types line up (in a very
hacky and one-off implementation, since the only tryable type for now is
the `Result<V, E>` type (though there's no reason why it has to be, it
could change when "interfaces" are added)), a match expression is
emitted which handles the logic. For example, the following code:
```
func foo(): Result<Int, String> = ...
func bar(): Result<Int, String> {
  val a = try foo()
  Result.Ok(a + 0.1)
}
```
After transformation, this gets emitted:
```
func foo(): Result<Int, String> = ...
func bar(): Result<Int, String> {
  val a = match foo() {
    Result.Ok(v) => v
    Result.Err e => return e
  }
  Result.Ok(a + 0.1)
}
```