The clang saga is complete!
Working on macos but building on ubuntu (via github actions) caused a
ton of confusion w.r.t. `gcc`/`clang`. Apparently on macos, `gcc` is
just a thin wrapper for `clang`. Most of the arguments are the same, but
the thing that kept throwing me off was the `-lgc` flag to link the
`gc.a` static library file within `transpile/target/c/libgc/lib`. What
worked on my local macos machine did not work when run against the
ubuntu environment in github actions, and sometimes vice versa. In the
end, I ended up switching to using `clang` (explicitly) and passing the
`gc.a` file as part of the source files, rather than relying on the `-L`
flag to provide the library path. I don't know exactly why this worked,
but I'm not questioning it. I also needed to explicitly link against the
`math` lib (with `-lm`) in order for it to succeed when building on
ubuntu, even though this appears to be unnecessary on macos (though it
doesn't hurt).