Strings & Arrays!
- Implemented Strings and Arrays in the compile-to-c implementation.
With this comes string and array literals, as well as string
concatenation via the `+` operator. Also, with the array implementation
comes the variadic functionality of the `println` function, which
allowed some cleanup in existing tests.
- This also adds the concept of "lifting" expressions during C code
generation. The idea here is that there is certain abra code that, in
order to be represented in terms of C code, cannot be represented in a
single expression. We previously saw this with if-expressions, but it
became apparent again in array literals (since array items might need to
be allocated on the heap at creation-time). So the "lifting"
functionality is essentially an ast walker which runs at the statement
level in a block, performing the necessary setup logic for the
underlying expression, and then saves a (randomly-generated) name which
refers to the result. When the actual code generator runs and sees the
node in question, it emits those variable names from the lifting step.