Add initial implementation of methods
- Add `is_bound` and `self` fields on the `callable_ctx` struct flavors.
A "method" is just an arity+1 function, with a `self` value stored in
the callable context. The `self` value will be passed in as the implicit
argument into the function implementation.
- Add some macro maddness in the `callable.h` file to less annoyingly
define all of the supported arities in our transpiler. For now, let's
just allow for arity-11 functions. Anything beyond that is kind of nuts
anyway (maybe this should be supported in the typechecker?). This limit
can always be easily expanded later.