Closure support
Support for functions which close over variables in their lexical
environment. All functions (even non-closure functions) now receive a
`value_t* env` as the first argument, which will contain the closed-over
values from its lexical scope. When a variable is referenced in such a
way, it is "lifted" up from its location on the stack to a location on
the heap; subsequent references to this lifted data must be redirected
in that way.