Support for implicit default param values
Supply the default value for the parameter if it was either explicitly/implicitly omitted.
An explicit omission involves emitting the passing of `None` values for
default-valued parameters which don't have a value provided. This is
possible for functions whose arity is known at compile-time. For example:
  func foo(i = 0): Int = i
  foo()
An implicit omission occurs when the function's arity cannot be known at
compile-time. Since no `None` values can be emitted during compilation,
the parameters must resolve to their default values dynamically at
runtime. This is implemented by requiring that every function accept an
`int8_t num_received_args` parameter, which will represent the number of
arguments passed to the function (this _is_ known at compile-time). At
runtime, the function will compare this value with each parameter and,
using the parameter's index, determine if the default value should be
used. For example:
  func foo(i = 0): Int = i
  func bar(fn: () => Int): Int = fn()
  bar(foo)