Maps!
Adds support for Map literals, including instantiation and the indexing
operator. This included implementation of the `value_hash` function for
all existing types (and support for future types using the assumption
that the hash method will always be at index 2 in a type's vtable).

Also, for some reason, I had forgotten to un-comment the tests for ||=
and &&= operators, as well as to _add_ tests for && and ||. So I did
that too.