Enum methods
- Also, fix latent bug with if/else-if expressions. The basic blocks
being backreferenced in the `phi` instructions were not being
correctly determined, since nested if-expressions (as is the case with
else-if nodes) generate additional basic blocks. It's _these_ extra
basic blocks that need to be referenced in the `phi` instruction,
otherwise llvm cannot properly resolve the chain of if/else-if/else
basic blocks. This is the issue that I encountered the other day when
trying to build out invocations for generic expressions as functions; I
should in theory, with this knowledge, be able to eliminate the (very
convoluted) `function_call` function.