Prepare for module compilation
Each module has an index number, based on the dependency order
determined by the module loader. For each module index, we generate an
initializer function (eg. for module index 0, we create `$0$init()`)
which is where the top-level logic within a module lives. All other
values are namespaced w.r.t. that module index, and the initializer for
each module is called within the topmost `__main` entrypoint.