Parsing prelude types' methods
Allow for `stub_mode` in the `Parser`, which will not fail when parsing
a function definition without a body, as is the case in
`prelude.stub.abra` (and other future potential `*.stub.abra` files).
With this in place, add `func isEmpty(self): Bool` methods to the 3
builtin types (`Array`, `Set`, and `Map`), and add tests for them.
The tests also allow for some much needed cleanup of the prelude_module
test, which is an added bonus.