C2: Moving from a pointer-based model
Previously, all values in the c2 implementation had been heap-allocated
pointers, even primitives (ints, floats, bools). This always felt icky
as it introduced additional overhead for arithmetic and simple values,
and it also added additional confusion for captured locals in closures
(which would have been represented as a heap-allocated pointer to a pointer
to a heap-allocated value...).
This new model is a bit riskier since it relies on assumptions about
struct field alignment, but basically every value is now a `size_t type_id`
and then some `data`. This `data` is always expected to be 8 bytes wide,
and for primitive values (int, float, bool) it will contain the value;
for heap-allocated values (string, array, tuple, etc) it will be a
pointer to that value (denoted with an `_Inner` suffix to the
parent struct name). This makes implementing captures in closures a bit
more straightforward, but it does introduce some potential confusion in
casting.
For example, when casting a string to an any type, we must use pointer
indirection (aka a reinterpret_cast), but the value we operate on must
be a non-temporary value. This is fine, because the compiler works in a
pseudo-SSA style (and any place that hadn't been has been updated in
this changeset to ensure no breakage).
Overall, I think this is a change for the better. It makes arithmetic
and boolean operations faster because it eliminates a pointer
dereference, and it also simplifies the mental model for dealing with
heap-allocated objects and closure captures.