LLVM2: Compiling array literals
Add support for instance and static method generation, to generate code
for constructing arrays. This uses the native implementation of
Array.withCapacity (which in turn leverages the built-in Pointer<T>
type). At this point, only the code which is actually reachable is
generated, so dead code removal is the default (I'm not sure if I'll
be able to keep that up, but for now it's pretty cool).

For now, tests will fail because I've modified the Array#push method to
not accept a varargs, to simplify the implementation. Perhaps I should
just modify the signature permanantly?