Generate toString methods
LLVM2: Add support for toString methods, and make existing builtin
toString implementations lazily-generated to fit the ongoing paradigm of
lazily-generated function definitions.
This change also fixes a bug where heap-allocated types (namely
`Array<T>`, which is what prompted the discovery of the bug) weren't
being properly allocated, leading to weird artifacts when getelementptr
was used.
Typechecker2: Allow user-defined toString methods to be defined in
types; if no method is defined, then a signature will be automatically
added for the type (and subsequently later generated in compilation).
Note that, for now, custom toString methods _MUST_ be the first instance
method defined for a type, otherwise certain assumptions fall apart.
This is especially apparent for generics, since the following code must
be able to typecheck and compile:
```
func echo<T>(value: T): String = value.toString()
```
The only way we can reliably know that any value of type `T` has a
`toString` method is if we can assume it's in a known position - index 0
in the type's `methods` vector. This will also eventually be true for
`eq` and `hash` methods but we're pretty far from that in the LLVM2
implementation.
In the future, it shouldn't matter which order the methods are defined
if there are custom implementations provided for toString/eq/hash; the
typechecker can shuffle them around prior to compilation if need be.