LLVM2: Optionals
Create Option<T> types; in order to accomplish this, we generate struct
definitions for each Option<T> of the shape:
```
type Option<T> {
  isSet: Bool
  value: T
}
```
This data structure is totally opaque and is only available at runtime;
for `None` values, `isSet` will be `false`, and `value` will just be
whatever the zero-value is for the underlying type (NULL for pointer
types, 0 for other types).

In order to achieve this, the typechecker needs to provide some
additonal data during typechecking. For example, consider the following
code:
```
val a = 1
val x: Int? = a
```
When `a` is declared, it's known to be of type `Int`, and the compiler
can emit correct instructions to treat it as such. However, when it's
used in place of an `Int?` type (as in the assignment on the second
line) we previously had no way of knowing whether to "upcast" the `Int`
to an `Int?`. Since different code is emitted in the compiler depending
on the value's use, we need to know this information. So the concept of
`resolved_type_id` is introduced, and is used to determine if any
transformations need to be done to the value after its creation. This
concept will be used in other situations in the future as well, for
things like `Result<V, E>` and also for assigning values to
trait/interface values.