Adding compilation for `Any`
Parser: Fix bug where `x < y` would be incorrectly treated as a type
parameter being passed to the `x` identifier.
Typechecker2: Add resolved_type_id value to some additional typed nodes.
Compiler2: Add compilation for "traits" (in reality, just the `Any`
type, as that's the only trait implementation at the moment). This is
done by leveraging wide pointers - a trait instance, eg Any,
is represented as a pointer to a vtable for the type of the given
underlying value, as well as a 64-bit integer which holds the underlying
value. If the value is a primitive type (Int, Float, Bool), the value is
bitcast to a 64-bit integer; if the value is a pointer-backed object,
the pointer is cast to a 64-bit integer. If the value is of type `Option<T>`,
then the value is first moved to the heap, and then treated as a pointer
value. The vtable is statically created for each type that's wrapped in
the given trait, and includes the typeid (todo: use auto-incrementing
type ids rather than the per-module, per-scope version that the
typechecker uses) and pointers to wrapper functions for the methods of
the underlying value's type. These wrapper functions accept a 64-bit
integer, perform the reverse transformation depending on the type, and
then call the real method with the properly-cast value. Using this
implementation, I was able to get `println` working (for all but one
case, which I think will require some typechecker modificiations; but
it's an edge case with many workarounds, and I think it can wait until
self-hosting to correct).