String concatenation
Typechecker2: Add `fn_type_id` to `Function` instances which eliminates a
lot of the complexity of determining function types after the fact
(especially in non-typechecker contexts, where mutations to `Project`
should be forbidden). Also, add `type_arg_ids` to `TypedNode::Invocation`;
since the type arguments to any function call are known (and/or inferred)
at typechecking-time, it doesn't make sense to also have to specify them
manually in the source code. Especially in the context of the new LLVM
compiler, where different functions are emitted for each generic usage.
Compiler2: Add support for the `+` operator to facilitate string
concatenation if either the LHS or RHS is a `String`.
Prelude: The prelude received the `String#concat<T>(suffix: T)` method
implemented. Note that this changes the signature of that method since
it previously had a varargs of `Any` after it; eventually we may add
this back, or perhaps come up with some improved signature for it. Also,
since all invocations now come with pre-determined `type_arg_ids`, we
can remove some places where we had previously needed to specify them.