Chore: refactoring resolved_generics
Refactor the notion of resolved generics into a struct with helper
methods. This makes it a lot easier to see patterns and to better convey
intent when building up the map of resolved generics. This also splits
the concept of generics out of the context_stack, which is helpful as
that will be used in future changesets to implement closures.