Function values
LLVM2: Add compilation for non-closure function values. This supports
creating higher- & lower-arity wrapper functions for function with
optional parameters as well as discarded parameters.
Typechecker2: Treat it as a typechecker error if a variadic function
is attempted to be assigned to another function value. This, while
technically doable with enough consideration, is an edge case that can
be ignored for now.