Nested closures
LLVM2: Allow for closures to be called statically more easily, and also
from a lexical scope other than the one in which they're defined. This
is accomplished by passing the closure's captures array as a capture
into functions which call that closure, effectively lifting that
function up into a closure as well.