Tuples
LLVM2: Add support for tuple literals and for indexing into tuples. This
also does some much-needed cleanup for casting an expression to a target
type (as we do for implicit optional conversion and trait (ie `Any`)
conversion). Additionally, this adds a convenient abstraction for
getting/compiling the `toString` implementation for a given `TypeId`.
This is useful in generating `toString` implementations for `Tuple`
types, but is also useful for printing out the value of the last
expression in the module being executed.