Number stuff
This adds changes across the board related to Ints and Int operators.
This includes hex/bin Int literals, the shift left and shift right
operators, bitwise and, or, and xor operators (which just use the same `&&`,
`||`, and `^` symbols as the boolean counterparts), and the implementation
of Pow in LLVM2.

Note: the VM, C transpilers (both v1 and v2), v1 LLVM compiler, and v1
typechecker do not implement these new shift left/right operators or the
bitwise operators.