Hash methods
Typechecker2: Add logic to append a dummy `hash` method to necessary
types if an implementation isn't provided; if one is provided by the
user, then we don't auto-generate one. Regardless, the `hash` method
will become method index 1 in the struct's methods array; I also updated
the existing `toString` logic to behave the same, except it places it at
method index 0. This solidifies some assumptions made earlier on, and
also allows `toString` and `hash` methods (and eventually `equals`; aka
the 3 methods on the future `Any` trait when traits are added) to be
defined in any location (previously `toString` _had_ to be first if a
custom implementation was provided, which was not great).
LLVM2: Add code generation for `hash` methods, as well as auto-generated
methods for some builtin types (`Int`, `Float`, `Bool`).
Prelude: Add `hash` implementation for `String` and `Array`. The
`String#hash` method uses a new intrinsic method, `Byte#asInt` which
requires native support from within the compiler.