Condition bindings and break/continue
This is working towards a Map#insert implementation, using that as a
forcing function to implement some todos.

Typechecker2: Add block terminators to If and While typed ast nodes.
These will be very helpful during compilation since we can use them to
change the way code is emitted during break/continue (and eventually
return, though that's not part of this commit).
LLVM2: Emit code for break/continue, using the terminators added to the
If and While typed ast nodes. Eventually there will need to be
terminators added to many more places, but for now this is good enough.
Also, add support for condition bindings in if- and while-conditions,
and allow those conditions to be Optional types as well (rather than
just Bool, as it was prior).