Maps
The motivating example was getting basic hashmap construction to work
using the `Map#insert` method. This required changes to how Option
type instances are flattened as well as several other things. The
majority of the work was done in the Prelude, with some supporting work
done as unimplemented/broken logic was hit in the compiler.