Optional-safe invocations & Set
Previously, only optional-safe accessors (ie. field accessors) were
supported, but this provides implementation for calling methods on an
Option value in an option-safe manner.

In addition, this changeset also provides the beginnings of the
implementation of the `Set` type. However, this change highlighted a
huge flaw in the current way generics are resolved at compile-time;
namely, since `Set<T>` uses `T` as its generic, any dependent types
which also have the same-named generic (for example, `Array<T>`, which
is used in `Map<K, V>` as its entries of `Array<MapEntry<K, V>>`...)
cannot be resolved correctly. In retrospect, I don't know why I didn't
realize this would be a problem sooner. At any rate, this will likely
require a fundamental change and since this changeset was already
getting a bit unruly I decided to just change `Set<T>` to `Set<S>` for
now, using `S` as the generic name rather than `T`. Not great, but it
works for now until I can fix this for real in the next commit.