Chore: coalesce implementations of some functions
LLVM2: Right now there are duplicates of several of the functions within the
compiler, namely the *_by_type and *_by_type_id functions. The reason I
had done the *_by_type functions were to support code generation for
adhoc Type instances, ie Types which were created within the compiler's
execution but did not have a corresponding TypeId in the Project as it
was not created during typecheck-time. This was really only convenient
in a handful of cases, which were easy to write other helper functions
for. This allowed me to get rid of the *_by_type functions in favor of
the *_by_type_id variants.