Major refactor to ResolvedGenerics
LLVM2: Because everything in the compiler is generated lazily (so as to
hopefully achieve dead-code removal by default), we need to keep track
of the values to resolve Generics to as we generate types and methods.
This gets complicated to manage, especially when the same method is
being generated but in a different context; for example, when generating
the `Map<K, V>#entries` method we generate a `Set<T = (K, V)>`, but
under the hood, `Set<T>` uses a `Map<K = T, V = Bool>`, which uses an
`Array<MapEntry<K = (K, V), V = Bool>?>`. In order to resolve this
complex dependency chain of generics, we need to keep track of
ResolvedGenerics in "layers", and partially resolve the types as new
layers are added. This leverages the fact that generated types are
indexed in the llvm module by a unique name identifier; storing this
identifier allows us to retrieve the generated type later on down the
chain when the original context is lost/overridden.

It's likely that there are still many quirks with this approach and it's
not perfect, but it's a work in progress and everything seems to work
for now.