Equality between objects
Typechecker2: Add a default `eq` method on all struct types
LLVM2: Add an autoincrementing, runtime `typeid` onto each object type.
This will be used later on to generate type-based `match` cases, but for
now is used as the first comparison in equality checks between objects.
If two values are of the same type (ie. they have the same `typeid`), we
call the `eq` method of the underlying type.