(Hopefully) fixing the resolved_generics issue once and for all
The issue with the way ResolvedGenerics previously worked is that we
lose initial context as we descent down into types' scopes; this is
especially apparent when types are self-referential (as in `Map#entries`,
which generates a `Set`, which in turn contains a `Map`). When
generating for that inner map, the value for the generic `K` is defined
in terms of the original top-level Map's `K` value, which is lost along
the way. For example:
  - Map<K = String, V = Int>#entries => returns Set<T = (K, V)>
    (where K = String and V = Int)
  - Set<T = (K, V)> has a field `_map: Map<K = T, V = Bool>`
    (where T = (K, V), but not the _current_ Map's `K` value; it's the original Map's `K` value)

This is further complicated by the way that TypeIds work, and the fact
that so much is based around TypeIds as a cheap key into the Project's
Types array. What we'd ideally like to have is a model where as we
descent through the heirarchies of type contexts during compilation, we
can create novel Types to build up the values and fill in the known
generics as we go. Since we can't modify the Project after the fact (or
rather, rust realy doesn't make it easy to pass around a mutable
reference in this way), we need to have an internal list of "ad hoc"
Types that are maintained within the Compiler itself. TypeIds with a
ScopeId of `ScopeId::BOGUS` are treated as sentinel values and will be
fetched from this "ad hoc" array rather from than the Project. This
meant replacing all occurrences of `project.get_type_by_id` (and other
similar methods, like `project.type_is_option`, etc) with wrapper
implementation which first check the ScopeId of the TypeId, and refer to
the appropriate Types array accordingly.