Early returns
Typechecker2: Add methods to determine if a TypedNode contains any
terminators (`break`, `continue`, or `return` statements), and whether
that node results in a `return` (for example, an if-statement in which
both branches contain a `return` would be treated as a "returning"
node).
LLVM2: Add code emitting logic for return statements.
Prelude: Clean up some code that no longer needs to be written in
awkward ways thanks to the inclusion of early returns.