For-loops
Typechecker2: Change for-loops for maps to be more in line with that of
arrays; rather than switching the key and the value, just allow the
iterator to provide instances of `(K, V)` and then the value can be
destructured in the loop binding.
LLVM2: Compilation for for-loops. Using an Array/Set/Map as the target
will perform an implicit invocation of the type's `#iterator()` method
which will produce an Iterator instance; calling `#next()` on that
instance will provide either the next value or `None`, indicating that
the loop is done. If instead of an Array/Set/Map, an instance of an
Iterator is provided, the for-loop will just call the `#next()` method
just as it would in the previous case. This allows for a lot of
flexibility in the ways that for-loops can be used (such as with the
builtin `range` function). Doing things this way is a lot more efficient
as it doesn't require an ahead-of-time array to be generated as with the
old `#enumerate()` idea.