Match statements/expressions
Typechecker2: There were quite a few changes required here - namely, I
wasn't even returning the typed match-case information in the typed
Match node, so it was impossible to generate code for it in the
compiler. Additionaly, the optional case binding variable needed to have
its `VarId` passed along in the typed Match node.
LLVM2: Compilation for match expressions/statements. This only supports
the basic cases kinds of `None` and wildcard for now, but it supports
block terminators in match expressions' case bodies, as well as properly
handling case bindings (at least, for the None/non-None case).