Extract out most intrinsics into std/_instrinsics.abra
Prelude: Import `std/_intrinsics.abra` to abstract away a lot of the
intrinsic implementations. There's still some that remain, but I'm okay
with this for now.
LLVM2: Fix small bug where a Unit-return function with an empty body
would fail to compile.
Typechecker2: This was a sizeable amount of work went - I had to
refactor the way in which modules are loaded in order to allow the
prelude module to effectively be treated like any other module, and
allow it to have an import of the `_intrinsics.abra` module. In doing
this, I changed the module resolution strategy to be file-based as
opposed to the admittedly nonsensical way I had been doing it before;
it's a lot more reasonable now I think, though I'm sure I'll change it
again in the future if I ever get around to removing the two competing
`ModuleId` implementations (`parser::ast::ModuleId` and
`typechecker::typechecker2::ModuleId`). Anyway, modules are now
correctly resolved with respect to other modules which import them, and
static modules (such as `prelude` or `io`) are resolved starting from
the `std/` directory (which is also configurable via a CLI flag).