Adding libc module (with @CBinding funcs)
The goal of this was initially to write a `readFile` function in a new
`fs` module, using only functions from libc (`open`, `lseek`, and
`read`). However, that turned into something greater as I wanted an easy
way to create c bindings in abra code; so as a proof of concept I
created the `write` function and replaced the `stdoutWrite` intrinsic
(which contained an implementation within the LLVM2 compiler) with just
a call to the `libc.write` function.
In order to do this, I needed to restructure some things around with
imports (which I missed in my earlier refactoring). This also
accidentally allowed me to move `Byte` and `Pointer`, two types which
consisted solely of `@Intrinsic` methods, into the
`std/_intrinsics.abra` module, which was a nice added benefit.

Parser: The change here is that the `ParseResult` now has an `imports`
vector which contains the actual import node. This effectively removes
the ImportNode from the AST (this isn't true for older versions, but it
is true for Typechecker2->LLVM2).
Typechecker2: Process all of the imports prior to doing any
function/struct/enum passes. This ensures that all imported types which
are referenced in function signatures are properly recognized. In doing
this, I removed TypedNode::Import entirely, as it's no longer part of
the AST and instead is just handled separately (thanks to changes in the
Parser's `ParseResult`). I also cache the `std/_intrinsics.abra` module
in the `Project`, since I need to check against it in the compiler to
preserve the name of `Byte` and `Pointer` (since those two types' names
are generated specially). Probably, at some point, I should figure out a
different way to do this, but whatever.
LLVM2: Remove the `stdout_write` intrinsic since I replaced it with a
native implementation using solely `libc.write`. Also, handle functions
annotated with `@CBinding` by constructing a wrapper function (this is
helpful for converting between native-c types and abra types (like Int
for example, which is 64-bit in abra but 32-bit in the libc functions
I've created bindings for so far) and for any other kind of things I may
want to do in the future, like checking errno perhaps).
Prelude/fs/libc/_intrinsics: Added libc module, moved `Byte` and `Pointer`
from `prelude` into `_intrinsics`, started `fs` module but commented it
out when I decided instead to focus on the `libc.write`
proof-of-concept.