Importing variables (globals)
The motivating example here is switching `libc.STDOUT_FILENO` from a
`func` to a `val`. It had previously been a function because importing a
variable and using it within a function in a separate module was not
supported because exported variables were not declared as globals and
were thus inaccessible (there was also another issue at play within the
typechecker).

Typechecker2: Fix the aforementioned issue with loading imported
variables. Imported variables are no longer re-declared as new variables
within the importing module but instead are just referenced directly.
This seems like one of those obvious things in hindsight.
Also, make sure that an imported variable that's used within a function
does _not_ get treated as a capture - this is important, since the
exported variable is a global and does not need to be captured. In
the future, I should probably just make all top-level variables globals
to really minimize on the amount of closures that are created.
LLVM2: Ensure that a module's imports are always compiled prior to
compiling the module itself (this also holds for the prelude module,
which can have its own imports now). Also, exported variables should now
be globals.
Prelude/libc: Change `libc.STDOUT_FILENO` from a function to a variable,
now that variables can be exported and used in importing modules'
functions (as is the case with `prelude.stdoutWrite`).