Abstract over different errno implementations
Operating systems have different implementations of `errno`, and
portability in C is obtained via importing `errno.h`. In this
implementation, we need to generate different code when obtaining the
`errno` value depending on the OS the binary is built for; right now I'm
only supporting macos (which uses `int *__error()`) and linux (which
uses `int *__erno_location()`). This is the first time I'm doing
OS-specific logic within the LLVM2 compiler, so if this pattern becomes
more common I'll likely have to come up with some better representation
of these differences.