Add `Process` builtin for args/env
Prelude: Add `Process.args()` and `Process.getEnvVar()`, two static
methods on a new `Process` type which allows access to the arguments
supplied to the program as well as any environment variables that may be
present. This required some new intrinsic/libc functionality.
LLVM2: Add backing code for new intrinsics, as well as code to be
able to pass program arguments into the executable. This also applies to
tests.
Typechecker2: Fix minor bug that only became apparent during
implementation of `Process.args()`; an empty Array/Set/Map would be
treated as a returning-terminator node and would cause no `ret`
instruction to be emitted.