Change the order of arguments to clang
For some reason, the order of the compilation targets passed to `clang`
matters, as when the `libgc.a` file is provided prior to the compiled
`.ll` file in the arguments list, it fails to find the necessary symbols
for compilation (namely `GC_init` and `GC_malloc`). Switching the order,
so that clang first sees the `.ll` module and then the `libgc.a` static
lib file seems to fix this issue. I'm sure there's some reason why this
is necessary, but I don't know it. I just hope I remember next time I
need this knowledge as this was quite frustrating to debug.